added enhancements based on feedback, reformat drawing class, new plot styles, init module file, detach matplotlib dependency for circuit
