+++ src/qibo/ui/__init__.py
@@ -1,0 +1,1 @@
+from qibo.ui.mpldrawer import plot
